fix: behaviour for groups also
